added confirm deleting car

+++ src/app/dashboard/cars/cars.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import swal from 'sweetalert2';
 
 import { Car } from '../../shared/car.model';
 import { CarsService } from '../../services/cars.service';
@@ -8,6 +9,7 @@ import { CarsService } from '../../services/cars.service';
 })
 export class CarsComponent implements OnInit {
   cars: Car[];
+  processing: boolean = false;
 
   constructor(private carsService: CarsService) {  }
 
@@ -18,8 +20,24 @@ export class CarsComponent implements OnInit {
   }
 
   delete(id: number): void {
-    this.carsService.delete(id).then(() => {
-      this.cars = this.cars.filter(item => item.id !== id);
+    swal({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      type: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#DD6B55",
+      confirmButtonText: "Yes, delete it!",
+    }).then(() => {
+      this.processing = true;
+      this.carsService.delete(id).then(() => {
+        this.processing = false;
+        this.cars = this.cars.filter(item => item.id !== id);
+        swal(
+          'Deleted!',
+          'Your car has been deleted.',
+          'success'
+        );
+      });
     });
   }
 }
